requirements.txt
annotated-types==0.7.0
numpy==1.26.4
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyYAML==6.0.2
scipy==1.15.2
sounddevice==0.5.1
typing-inspection==0.4.0
typing_extensions==4.13.2
